+++ lib/SimpleCacheAPCu.php
@@ -405,8 +405,6 @@ class SimpleCacheAPCu implements \Psr\SimpleCache\CacheInterface
      *                                             serializable.
      * @param null|int|string|\DateInterval $ttl   (optional) The TTL value of this item.
      *
-     * @psalm-suppress MoreSpecificImplementedParamType
-     *
      * @return bool True on success and false on failure.
      */
     public function set($key, $value, $ttl = null): bool
@@ -525,8 +523,6 @@ class SimpleCacheAPCu implements \Psr\SimpleCache\CacheInterface
      *
      * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
      * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @psalm-suppress MoreSpecificImplementedParamType
      *
      * @return bool True on success and false on failure.
      */